feat: users rating list

+++ modules/services/api-empireapi/src/commonMain/kotlin/ru/astrainteractive/empireapi/models/response/GenericPagedModel.kt
@@ -7,5 +7,5 @@ class GenericPagedModel<T>(
     override val data: List<T>,
     override val total: Long,
     override val currentPageAmount: Int,
-    override val page: Long
+    override val page: Int
 ) : PagedModel<T>
