+++ modules/services/api-empireapi/src/commonMain/kotlin/ru/astrainteractive/empireapi/models/response/PagedModel.kt
@@ -4,5 +4,5 @@ interface PagedModel<T> {
     val data: List<T>
     val total: Long
     val currentPageAmount: Int
-    val page: Long
+    val page: Int
 }
